functions/requirements.txt
fastapi
google-cloud-language
google-cloud-storage
requests
